slirp: fix Win64 structure issues
Fix three x86-64 compatibility issues:
- Bitfield layout: struct ip and tcphdr use u_int8_t nibbles
- SIZEOF_CHAR_P: set to 8 (was hardcoded to 4)
- Address space: link at 0x400000 with IMAGE_FILE_LARGE_ADDRESS_AWARE
  cleared to confine to 32 bits

Add _Static_asserts for structure sizes and -t self-test mode (ARP/DHCP)
to verify wire format parsing before guest use.

+++ src/colinux/user/slirp/tcp.h
@@ -53,16 +53,17 @@ extern struct socket *tcp_last_so;
  * TCP header.
  * Per RFC 793, September, 1981.
  */
+/* u_int8_t for the nibble pair, not u_int -- see struct ip in ip.h. */
 struct tcphdr {
 	u_int16_t	th_sport;		/* source port */
 	u_int16_t	th_dport;		/* destination port */
 	tcp_seq	th_seq;			/* sequence number */
 	tcp_seq	th_ack;			/* acknowledgement number */
 #ifdef WORDS_BIGENDIAN
-	u_int	th_off:4,		/* data offset */
+	u_int8_t th_off:4,		/* data offset */
 		th_x2:4;		/* (unused) */
 #else
-	u_int	th_x2:4,		/* (unused) */
+	u_int8_t th_x2:4,		/* (unused) */
 		th_off:4;		/* data offset */
 #endif
 	u_int8_t	th_flags;
